work on mouse cursol

+++ lib.h
@@ -5,7 +5,7 @@
 
 void putc(CHAR16 c);
 void puts(CHAR16 *s);
-void putxval(UINT64 val, UINT64 column);
+void putxval(UINT64 val, UINT8 column);
 CHAR16 getc(void);
 UINT32 gets(CHAR16 *buf, UINT32 buf_size);
 int strcmp(const CHAR16 *s1, const CHAR16 *s2);
